creacion del componente register y correccion de componente login
