+++ client/src/App.js
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Catalogue from './components/Catalogue/Catalogue';
-import Login from './components/Login/Login'
+import UserConnection from './components/UserConnection/UserConnection'
 import Home from './components/Home/Home';
 import Navbar from './components/Navbar/Navbar';
 
@@ -10,9 +10,10 @@ function App() {
     <>
       <Navbar />
       <Routes>
-        <Route path='/' element={<Home></Home>} ></Route>
-        <Route path='catalogue' element={<Catalogue></Catalogue>} ></Route>
-        <Route path='login' element={<Login />} ></Route>
+        <Route path='/' element={<Home />} />
+        <Route path='catalogue' element={<Catalogue />} />
+        <Route path='login' element={<UserConnection />} />
+        <Route path='register' element={<UserConnection isLogin={false} />} />
       </Routes>
     </>
   );
