+++ client/src/components/UserConnection/UserConnection.css
@@ -1,4 +1,4 @@
 .width {
     width: 50vw;
-    height: 85vh;
+    height: 100vh;
 }
